% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialetti.R
\name{get_napoletano}
\alias{get_napoletano}
\title{get_napoletano}
\usage{
get_napoletano(txt)
}
\arguments{
\item{txt}{provide a text you want to translate from Iatlian into Napoletano}
}
\description{
Translates Italian to Napoletano
}
